pyproject.toml
[project]
name = "customer-financial-health-analyzer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fpdf>=1.7.2",
    "matplotlib>=3.10.1",
    "mlflow>=2.21.3",
    "ollama>=0.4.7",
    "pandas>=2.2.3",
    "pdfplumber>=0.11.6",
    "prefect>=3.3.4",
    "python-dateutil>=2.9.0.post0",
    "pyyaml>=6.0.2",
    "requests>=2.32.3",
    "ruff>=0.11.5",
]

[tool.ruff]
select = ["ALL"]  # Enables all rules
